feat: 클라우드DB 연결 및 서버 오류 해결

+++ backend/src/main/java/com/ssg9th2team/geharbang/domain/room/repository/jpa/RoomJpaRepository.java
@@ -3,6 +3,5 @@ package com.ssg9th2team.geharbang.domain.room.repository.jpa;
 import com.ssg9th2team.geharbang.domain.accommodation.entity.Accommodation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-
-public interface RoomJpaRepository extends JpaRepository<Accommodation, Integer> {
+public interface RoomJpaRepository extends JpaRepository<Accommodation, Long> {
 }
